# feat: generate CPU & PGO profiles