## Proposed changes

* Close #6057
* Close #5691
* Close #6060
* Close #6061 

## Proof of Concept

The following steps demonstrate the process of compiling and benchmarking the `nuclei` binary with and w/o PGO.

### Steps  

1. Compile the test binary and run the benchmark:
   ```bash
   make build-test
   DISABLE_STDOUT=1 ./bin/nuclei.test -test.run - -test.bench=. -test.benchmem -test.count=10 ./cmd/nuclei/ | tee nopgo.txt
   ```
   
2. Generate CPU profiles:
   ```bash
   go run cmd/nuclei/main.go -u "https://scanme.sh" -profile-mem="nuclei" -no-mhe -silent >/dev/null
   mv nuclei.cpu cmd/nuclei/default.pgo
   ```
   Alternatively, extract this PGO file ([pgo.zip](https://github.com/user-attachments/files/18866745/pgo.zip)) into the `cmd/nuclei/` directory - Duration: 95.70s, Total samples = 45.96s (48.03%).

3. Recompile with PGO and rerun the benchmark:
   ```bash
   make build-test
   DISABLE_STDOUT=1 ./bin/nuclei.test -test.run - -test.bench=. -test.benchmem -test.count=10 ./cmd/nuclei/ | tee withpgo.txt
   ```

### Benchmark Results  

```console
$ benchstat nopgo.txt withpgo.txt 
goos: linux
goarch: amd64
pkg: github.com/projectdiscovery/nuclei/v3/cmd/nuclei
cpu: 11th Gen Intel(R) Core(TM) i9-11900H @ 2.50GHz
                  │ nopgo.txt  │         withpgo.txt          │
                  │   sec/op   │   sec/op    vs base          │
RunEnumeration-16   183.0 ± 1%   183.0 ± 1%  ~ (p=0.105 n=10)

                  │  nopgo.txt   │             withpgo.txt             │
                  │     B/op     │     B/op      vs base               │
RunEnumeration-16   3.215Gi ± 0%   3.205Gi ± 0%  -0.32% (p=0.001 n=10)

                  │  nopgo.txt  │            withpgo.txt             │
                  │  allocs/op  │  allocs/op   vs base               │
RunEnumeration-16   33.94M ± 0%   33.83M ± 0%  -0.33% (p=0.001 n=10)
```

This suggests that PGO can help optimize memory efficiency, although the improvements might not always be dramatic. The effectiveness of PGO largely depends on the quality of the CPU samples used, which can vary based on the workload and profiling data. However, I'm still exploring ways to gather more comprehensive CPU samples - potentially by adjusting `types.Options` or expanding the number of `Targets` to improve sampling accuracy. The PGO data in this example was generated using the default `types.Options` from the CLI with only a single dummy server as the target, which may not fully capture real-world performance patterns.

## Checklist

<!-- Put an "x" in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code. -->

- [ ] Pull request is created against the [dev](https://github.com/projectdiscovery/nuclei/tree/dev) branch
- [ ] All checks passed (lint, unit/integration/regression tests etc.) with my changes
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
	- Introduced CPU profiling alongside memory profiling for enhanced performance insights.
	- Added functionality to suppress standard output based on an environment variable for improved configurability.
- **Documentation**
	- Updated instructions now cover both CPU and memory profiling, including how to visualize these profiles.
- **Build Enhancements**
	- Refined build configurations improve reproducibility and optimize the linking process.
- **Chores**
	- Enhanced version control settings exclude additional profiling files to maintain a cleaner repository.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->